add docker build (#1832)
* add docker build

* fixes

* lint

* fix perf issue

* nit

---------
