+++ pkg/docker/compose.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/ava-labs/avalanche-cli/pkg/constants"
 	"github.com/ava-labs/avalanche-cli/pkg/models"
-	"github.com/ava-labs/avalanche-cli/pkg/remoteconfig"
 	"github.com/ava-labs/avalanche-cli/pkg/utils"
 	"github.com/ava-labs/avalanche-cli/pkg/ux"
 )
@@ -59,7 +58,7 @@ func pushComposeFile(host *models.Host, localFile string, remoteFile string, mer
 	if fileExists && merge {
 		// upload new and merge files
 		ux.Logger.Info("Merging compose files")
-		tmpFile, err := host.CreateTemp()
+		tmpFile, err := host.CreateTempFile()
 		if err != nil {
 			return err
 		}
@@ -262,203 +261,4 @@ func HasRemoteComposeService(host *models.Host, composeFile string, service stri
 		}
 	}
 	return found, nil
-}
-
-// ValidateComposeFile validates a docker-compose file on a remote host.
-func ValidateComposeFile(host *models.Host, composeFile string, timeout time.Duration) error {
-	if output, err := host.Command(fmt.Sprintf("docker compose -f %s config", composeFile), nil, timeout); err != nil {
-		return fmt.Errorf("%w: %s", err, string(output))
-	}
-	return nil
-}
-
-func prepareAvalanchegoConfig(host *models.Host, networkID string) (string, string, error) {
-	avagoConf := remoteconfig.DefaultCliAvalancheConfig(host.IP, networkID)
-	nodeConf, err := remoteconfig.RenderAvalancheNodeConfig(avagoConf)
-	if err != nil {
-		return "", "", err
-	}
-	nodeConfFile, err := os.CreateTemp("", "avalanchecli-node-*.yml")
-	if err != nil {
-		return "", "", err
-	}
-	if err := os.WriteFile(nodeConfFile.Name(), nodeConf, constants.WriteReadUserOnlyPerms); err != nil {
-		return "", "", err
-	}
-	cChainConf, err := remoteconfig.RenderAvalancheCChainConfig(avagoConf)
-	if err != nil {
-		return "", "", err
-	}
-	cChainConfFile, err := os.CreateTemp("", "avalanchecli-cchain-*.yml")
-	if err != nil {
-		return "", "", err
-	}
-	if err := os.WriteFile(cChainConfFile.Name(), cChainConf, constants.WriteReadUserOnlyPerms); err != nil {
-		return "", "", err
-	}
-	return nodeConfFile.Name(), cChainConfFile.Name(), nil
-}
-
-// ComposeSSHSetupNode sets up an AvalancheGo node and dependencies on a remote host over SSH.
-func ComposeSSHSetupNode(host *models.Host, network models.Network, avalancheGoVersion string, withMonitoring bool) error {
-	for _, dir := range remoteconfig.RemoteFoldersToCreateAvalanchego() {
-		if err := host.MkdirAll(dir, constants.SSHFileOpsTimeout); err != nil {
-			return fmt.Errorf("failed to create directory %s: %w", dir, err)
-		}
-	}
-	ux.Logger.Info("avalancheCLI folder structure created on remote host %s", remoteconfig.RemoteFoldersToCreateAvalanchego())
-	// configs
-	networkID := network.NetworkIDFlagValue()
-	if network.Kind == models.Local || network.Kind == models.Devnet {
-		networkID = fmt.Sprintf("%d", network.ID)
-	}
-
-	nodeConfFile, cChainConfFile, err := prepareAvalanchegoConfig(host, networkID)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err := os.Remove(nodeConfFile); err != nil {
-			ux.Logger.Error("Error removing temporary file %s: %s", nodeConfFile, err)
-		}
-		if err := os.Remove(cChainConfFile); err != nil {
-			ux.Logger.Error("Error removing temporary file %s: %s", cChainConfFile, err)
-		}
-	}()
-
-	if err := host.Upload(nodeConfFile, remoteconfig.GetRemoteAvalancheNodeConfig(), constants.SSHFileOpsTimeout); err != nil {
-		return err
-	}
-	if err := host.Upload(cChainConfFile, remoteconfig.GetRemoteAvalancheCChainConfig(), constants.SSHFileOpsTimeout); err != nil {
-		return err
-	}
-	return ComposeOverSSH("Compose Node",
-		host,
-		constants.SSHScriptTimeout,
-		"templates/avalanchego.docker-compose.yml",
-		dockerComposeInputs{
-			AvalanchegoVersion: avalancheGoVersion,
-			WithMonitoring:     withMonitoring,
-			WithAvalanchego:    true,
-		})
-}
-
-func ComposeSSHSetupLoadTest(host *models.Host) error {
-	return ComposeOverSSH("Compose Node",
-		host,
-		constants.SSHScriptTimeout,
-		"templates/avalanchego.docker-compose.yml",
-		dockerComposeInputs{
-			WithMonitoring:  true,
-			WithAvalanchego: false,
-		})
-}
-
-// WasNodeSetupWithMonitoring checks if an AvalancheGo node was setup with monitoring on a remote host.
-func WasNodeSetupWithMonitoring(host *models.Host) (bool, error) {
-	return HasRemoteComposeService(host, utils.GetRemoteComposeFile(), "promtail", constants.SSHScriptTimeout)
-}
-
-func prepareGrafanaConfig() (string, string, string, string, error) {
-	grafanaDataSource, err := remoteconfig.RenderGrafanaLokiDataSourceConfig()
-	if err != nil {
-		return "", "", "", "", err
-	}
-	grafanaDataSourceFile, err := os.CreateTemp("", "avalanchecli-grafana-datasource-*.yml")
-	if err != nil {
-		return "", "", "", "", err
-	}
-	if err := os.WriteFile(grafanaDataSourceFile.Name(), grafanaDataSource, constants.WriteReadUserOnlyPerms); err != nil {
-		return "", "", "", "", err
-	}
-
-	grafanaPromDataSource, err := remoteconfig.RenderGrafanaPrometheusDataSourceConfigg()
-	if err != nil {
-		return "", "", "", "", err
-	}
-	grafanaPromDataSourceFile, err := os.CreateTemp("", "avalanchecli-grafana-prom-datasource-*.yml")
-	if err != nil {
-		return "", "", "", "", err
-	}
-	if err := os.WriteFile(grafanaPromDataSourceFile.Name(), grafanaPromDataSource, constants.WriteReadUserOnlyPerms); err != nil {
-		return "", "", "", "", err
-	}
-
-	grafanaDashboards, err := remoteconfig.RenderGrafanaDashboardConfig()
-	if err != nil {
-		return "", "", "", "", err
-	}
-	grafanaDashboardsFile, err := os.CreateTemp("", "avalanchecli-grafana-dashboards-*.yml")
-	if err != nil {
-		return "", "", "", "", err
-	}
-	if err := os.WriteFile(grafanaDashboardsFile.Name(), grafanaDashboards, constants.WriteReadUserOnlyPerms); err != nil {
-		return "", "", "", "", err
-	}
-
-	grafanaConfig, err := remoteconfig.RenderGrafanaConfig()
-	if err != nil {
-		return "", "", "", "", err
-	}
-	grafanaConfigFile, err := os.CreateTemp("", "avalanchecli-grafana-config-*.ini")
-	if err != nil {
-		return "", "", "", "", err
-	}
-	if err := os.WriteFile(grafanaConfigFile.Name(), grafanaConfig, constants.WriteReadUserOnlyPerms); err != nil {
-		return "", "", "", "", err
-	}
-	return grafanaConfigFile.Name(), grafanaDashboardsFile.Name(), grafanaDataSourceFile.Name(), grafanaPromDataSourceFile.Name(), nil
-}
-
-// ComposeSSHSetupMonitoring sets up monitoring using docker-compose.
-func ComposeSSHSetupMonitoring(host *models.Host) error {
-	grafanaConfigFile, grafanaDashboardsFile, grafanaLokiDatasourceFile, grafanaPromDatasourceFile, err := prepareGrafanaConfig()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err := os.Remove(grafanaLokiDatasourceFile); err != nil {
-			ux.Logger.Error("Error removing temporary file %s: %s", grafanaLokiDatasourceFile, err)
-		}
-		if err := os.Remove(grafanaPromDatasourceFile); err != nil {
-			ux.Logger.Error("Error removing temporary file %s: %s", grafanaPromDatasourceFile, err)
-		}
-		if err := os.Remove(grafanaDashboardsFile); err != nil {
-			ux.Logger.Error("Error removing temporary file %s: %s", grafanaDashboardsFile, err)
-		}
-		if err := os.Remove(grafanaConfigFile); err != nil {
-			ux.Logger.Error("Error removing temporary file %s: %s", grafanaConfigFile, err)
-		}
-	}()
-
-	grafanaLokiDatasourceRemoteFileName := filepath.Join(utils.GetRemoteComposeServicePath("grafana", "provisioning", "datasources"), "loki.yml")
-	if err := host.Upload(grafanaLokiDatasourceFile, grafanaLokiDatasourceRemoteFileName, constants.SSHFileOpsTimeout); err != nil {
-		return err
-	}
-	grafanaPromDatasourceFileName := filepath.Join(utils.GetRemoteComposeServicePath("grafana", "provisioning", "datasources"), "prometheus.yml")
-	if err := host.Upload(grafanaPromDatasourceFile, grafanaPromDatasourceFileName, constants.SSHFileOpsTimeout); err != nil {
-		return err
-	}
-	grafanaDashboardsRemoteFileName := filepath.Join(utils.GetRemoteComposeServicePath("grafana", "provisioning", "dashboards"), "dashboards.yml")
-	if err := host.Upload(grafanaDashboardsFile, grafanaDashboardsRemoteFileName, constants.SSHFileOpsTimeout); err != nil {
-		return err
-	}
-	grafanaConfigRemoteFileName := filepath.Join(utils.GetRemoteComposeServicePath("grafana"), "grafana.ini")
-	if err := host.Upload(grafanaConfigFile, grafanaConfigRemoteFileName, constants.SSHFileOpsTimeout); err != nil {
-		return err
-	}
-
-	return ComposeOverSSH("Setup Monitoring",
-		host,
-		constants.SSHScriptTimeout,
-		"templates/monitoring.docker-compose.yml",
-		dockerComposeInputs{})
-}
-
-func ComposeSSHSetupAWMRelayer(host *models.Host) error {
-	return ComposeOverSSH("Setup AWM Relayer",
-		host,
-		constants.SSHScriptTimeout,
-		"templates/awmrelayer.docker-compose.yml",
-		dockerComposeInputs{})
 }
